requirements/app.txt
galaxy.plugin.api==0.64
certifi==2019.3.9
beautifulsoup4==4.9.3
